Minor final fixups and seprated visit detail component
