+++ packages/esm-patient-chart-app/src/visit/visits-widget/past-visits-components/tests-summary.component.tsx
@@ -4,18 +4,7 @@ import styles from '../visit-detail-overview.scss';
 
 const TestsSummary = ({ tests }) => {
   const { t } = useTranslation();
-  return tests.length ? (
-    tests.map((test) => (
-      <>
-        <p>
-          <b>{test.testName}</b>
-        </p>
-        <p>{test.value}</p>
-      </>
-    ))
-  ) : (
-    <p className={styles.bodyLong01}>{t('noTestsFound', 'No Tests found.')}</p>
-  );
+  return <p className={styles.bodyLong01}>{t('noTestsFound', 'No tests found')}</p>;
 };
 
 export default TestsSummary;
