+++ packages/esm-patient-chart-app/src/visit/visits-widget/visit-detail-overview.component.tsx
@@ -1,69 +1,8 @@
-import React, { useEffect, useState, useMemo } from 'react';
+import React, { useEffect, useState } from 'react';
 import styles from './visit-detail-overview.scss';
 import { Visit, getVisitsForPatient, createErrorHandler } from '@openmrs/esm-framework';
-import Button from 'carbon-components-react/es/components/Button';
-import { useTranslation } from 'react-i18next';
-import dayjs from 'dayjs';
 import SkeletonText from 'carbon-components-react/es/components/SkeletonText';
-import EncounterListDataTable from './past-visits-components/encounter-list.component';
-import VisitSummary from './past-visits-components/visit-summary.component';
-import { Encounter } from './visit.resource';
-
-function formatDateTime(date) {
-  return dayjs(date).format('MMM DD, YYYY - hh:mm');
-}
-
-interface VisitDetailComponentProps {
-  visit: Visit;
-}
-
-const VisitDetailComponent: React.FC<VisitDetailComponentProps> = ({ visit }) => {
-  const { t } = useTranslation();
-  const [listView, setView] = useState<boolean>(true);
-  const encounters = useMemo(
-    () =>
-      visit.encounters.map((encounter: Encounter) => ({
-        id: encounter.uuid,
-        time: dayjs(encounter.encounterDateTime).format('hh:mm'),
-        encounterType: encounter.encounterType.display,
-        provider: encounter.encounterProviders.length > 0 ? encounter.encounterProviders[0].display : '',
-        obs: encounter.obs,
-      })),
-    [visit],
-  );
-
-  return (
-    <div className={styles.visitsDetailWidgetContainer}>
-      <div className={styles.visitsDetailHeaderContainer}>
-        <h4 className={styles.productiveHeading02}>
-          {visit.visitType.display}
-          <br />
-          <p className={styles.bodyLong01}>{formatDateTime(visit.startDatetime)}</p>
-        </h4>
-        <div className={styles.toggleButtons}>
-          <Button
-            className={`${styles.toggle} ${listView ? styles.toggleActive : ''}`}
-            size="small"
-            kind="ghost"
-            onClick={() => setView(true)}>
-            {t('allEncounters', 'All Encounters')}
-          </Button>
-          <Button
-            className={`${styles.toggle} ${!listView ? styles.toggleActive : ''}`}
-            size="small"
-            kind="ghost"
-            onClick={() => setView(false)}>
-            {t('visitSummary', 'Visit Summary')}
-          </Button>
-        </div>
-      </div>
-      {listView && visit?.encounters && <EncounterListDataTable visitUuid={visit.uuid} encounters={encounters} />}
-      {!listView && <VisitSummary encounters={visit.encounters} />}
-    </div>
-  );
-};
-
-// Base VisitsOverviewComponent
+import VisitDetailComponent from './past-visits-components/visit-detail.component';
 
 interface VisitOverviewComponentProps {
   patientUuid: string;
