+++ packages/esm-patient-chart-app/translations/en.json
@@ -18,6 +18,7 @@
   "Medications": "Medications",
   "newVisit": "New visit",
   "no": "No",
+  "noDiagnosesFound": "No Diagnoses found",
   "noEncountersFound": "No encounters found.",
   "noMedicationsFound": "No Medications found",
   "noNotesFound": "No notes found.",
@@ -33,6 +34,7 @@
   "startVisit": "Start visit",
   "startVisitPrompt": "No active visit is selected. Do you want to start a visit?",
   "Tests": "Tests",
+  "thereIsNoInformationToDisplayHere": "There is no information to display here",
   "typeOfVisit": "Type of visit",
   "visitDashboard": "Visit dashboard",
   "visitEndDate": "Visit end date",
